+++ static/traces/spanmetricsprocessor/internal/metadata/generated_status.go
